@@ -1,11 +1,12 @@
 {
   "name": "typeof-properties",
   "version": "2.0.0",
-  "description": "Validate the types of object properties.",
-  "main": "src/index.js",
+  "description": "Validate the type of object's properties.",
+  "main": "dist/typeof-properties.node.js",
   "scripts": {
-    "test": "node tests/jasmine.conf.js",
-    "build": "webpack --config webpack.config.js"
+    "test": "babel-node jasmine.conf.js",
+    "start": "webpack --config webpack.dev.js",
+    "build": "webpack --config webpack.prod.js"
   },
   "keywords": [
     "typeof",
@@ -22,7 +23,7 @@
   "author": "Paweł Rafałko",
   "license": "MIT",
   "dependencies": {
-    "of-type": "^2.2.0"
+    "of-type": "^3.0.0"
   },
   "homepage": "https://github.com/devrafalko/typeof-properties#readme",
   "bugs": {
@@ -33,13 +34,17 @@
     "url": "git+https://github.com/devrafalko/typeof-properties.git"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-loader": "^7.1.5",
-    "babel-preset-env": "^1.7.0",
-    "jasmine": "^3.1.0",
+    "@babel/cli": "^7.1.0",
+    "@babel/core": "^7.0.1",
+    "@babel/node": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "babel-loader": "^8.0.2",
+    "jasmine": "^3.3.1",
     "jasmine-spec-reporter": "^4.2.1",
-    "uglifyjs-webpack-plugin": "^1.2.7",
-    "webpack": "^4.16.1",
-    "webpack-cli": "^3.1.0"
+    "uglifyjs-webpack-plugin": "^1.3.0",
+    "webpack": "^4.19.1",
+    "webpack-cli": "^3.1.1",
+    "webpack-merge": "^4.1.4",
+    "webpack-node-externals": "^1.7.2"
   }
 }
