@@ -1,13 +1,11 @@
 /* global describe, it, expect */
-const path = require('path');
-const type = require(path.resolve('./src/index.js'));
-const scenario = require('./scenarios-arguments.js');
+import scenario from './data/scenarios-arguments.js';
 
 describe('When the module function type() is called', function () {
   describe('with callback defined and the arguments', function () {
     for (let i in scenario) {
       it(`type(${scenario[i].oActual},${scenario[i].oExpected}), it should throw TypeError with message "${scenario[i].message}"`, function () {
-        let binded = type.bind(this, scenario[i].actual, scenario[i].expected, () => { });
+        let binded = this.type.bind(this, scenario[i].actual, scenario[i].expected, () => { });
         expect(binded).toThrowError(TypeError, scenario[i].message);
       });
     }
@@ -15,7 +13,7 @@ describe('When the module function type() is called', function () {
   describe('without callback defined and the arguments', function () {
     for (let i in scenario) {
       it(`type(${scenario[i].oActual},${scenario[i].oExpected}), it should throw TypeError with message "${scenario[i].message}"`, function () {
-        let binded = type.bind(this, scenario[i].actual, scenario[i].expected, () => { });
+        let binded = this.type.bind(this, scenario[i].actual, scenario[i].expected, () => { });
         expect(binded).toThrowError(TypeError, scenario[i].message);
       });
     }
