@@ -1,13 +1,11 @@
 /* global jasmine, describe, it, expect */
-const path = require('path');
-const type = require(path.resolve('./src/index.js'));
-const scenario = require('./scenarios-invalid.js');
+import scenario from './data/scenarios-invalid.js';
 
 describe('The module function', function () {
   describe('when callback is not defined, it should throw TypeError with message', function () {
     for (let i in scenario) {
       it(`"${scenario[i].oMessage}"`, function () {
-        let binded = type.bind(this, scenario[i].actual, scenario[i].expected);
+        let binded = this.type.bind(this, scenario[i].actual, scenario[i].expected);
         expect(binded).toThrowError(TypeError, scenario[i].oMessage);
       });
     }
@@ -16,8 +14,8 @@ describe('The module function', function () {
     for (let i in scenario) {
       it(`{actual:${scenario[i].oActual},expected:${scenario[i].oExpected},name:${scenario[i].oName},message:${scenario[i].oMessage}}`, function () {
         let clb = jasmine.createSpy('clb');
-        let binded = type.bind(this, scenario[i].actual, scenario[i].expected, clb);
-        expect(type(scenario[i].actual, scenario[i].expected, () => { })).toBe(false);
+        let binded = this.type.bind(this, scenario[i].actual, scenario[i].expected, clb);
+        expect(this.type(scenario[i].actual, scenario[i].expected, () => { })).toBe(false);
         expect(binded).not.toThrowError();
         expect(clb).toHaveBeenCalledTimes(1);
         expect(clb).toHaveBeenCalledWith(jasmine.objectContaining({
